Move MutableExpr to synthesis
