+++ src/main/scala/leon/codegen/CodeGeneration.scala
@@ -1195,7 +1195,7 @@ trait CodeGeneration {
         mkBranch(b, al, fl, ch, canDelegateToMkExpr = false)
         ch << Label(fl) << POP << Ldc(0) << Label(al)
 
-      case MutableExpr(e) =>
+      case synthesis.utils.MutableExpr(e) =>
         mkExpr(e, ch)
 
       case _ => throw CompilationException("Unsupported expr " + e + " : " + e.getClass)
