+++ src/main/scala/leon/evaluators/RecursiveEvaluator.scala
@@ -831,7 +831,7 @@ abstract class RecursiveEvaluator(ctx: LeonContext, prog: Program, val bank: Eva
           throw RuntimeError("MatchError: "+rscrut.asString+" did not match any of the cases:\n"+cases)
       }
 
-    case MutableExpr(ex) =>
+    case synthesis.utils.MutableExpr(ex) =>
       e(ex)
 
     case gl: GenericValue => gl
