+++ src/main/scala/leon/solvers/z3/AbstractZ3Solver.scala
@@ -552,7 +552,7 @@ trait AbstractZ3Solver extends Z3Solver {
         val constructor = constructors.toB(tp)
         constructor(rec(InfiniteIntegerLiteral(id)))
 
-      case MutableExpr(ex) =>
+      case synthesis.utils.MutableExpr(ex) =>
         rec(ex)
 
       case other =>
